# src/main/resources/openapi/contratacion-personal.yaml
openapi: 3.0.1
info:
  title: Microservicio de Contratación de Personal
  version: 1.0.3
  description: >
    API para gestionar el reclutamiento y contratación de trabajadores.
    Incluye datos personales, familiares, educativos, de contacto y de seguridad social.
servers:
  - url: http://localhost:8080/api/contratacion
paths:
...
  /postulantes/{id}:
    get:
      summary: Obtener un postulante por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Datos del postulante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostulanteResponseDto'
        '404':
          description: Postulante no encontrado
    put:
      summary: Actualizar un postulante
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostulanteRequestDto'
      responses:
        '200':
          description: Postulante actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostulanteResponseDto'
    delete:
      summary: Eliminar un postulante
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Postulante eliminado

  # Catálogos
...
components:
  schemas:
...
    PostulanteRequestDto:
      type: object
      required:
        - nombres
        - apellidos
        - tipoDocumentoId
        - numeroDocumento
        - fechaNacimiento
        - sexoId
        - nacionalidad
        - direccion
      properties:
        nombres:
          type: string
        apellidos:
          type: string
        tipoDocumentoId:
          type: integer
        numeroDocumento:
          type: string
        fechaNacimiento:
          type: string
          format: date
        estadoCivilId:
          type: integer
        sexoId:
          type: integer
        nacionalidad:
          type: string
        telefono:
          type: string
        celular:
          type: string
        correo:
          type: string
        direccion:
          $ref: '#/components/schemas/DireccionDto'
        sistemaPrevisionalId:
          type: integer
        afpId:
          type: integer
        numeroSeguro:
          type: string
        gradoInstruccionId:
          type: integer
        institucionEducativa:
          type: string
        nombreConyuge:
          type: string
        numeroHijos:
          type: integer
        hijos:
          type: array
          items:
            $ref: '#/components/schemas/HijoDto'
    PostulanteResponseDto:
      allOf:
        - $ref: '#/components/schemas/PostulanteRequestDto'
        - type: object
          properties:
            id:
              type: integer
            fechaRegistro:
              type: string
              format: date-time

    # 📌 Submodelos
    HijoDto:
      type: object
      properties:
        nombre:
          type: string
        edad:
          type: integer
    DireccionDto:
      type: object
      properties:
        direccion:
          type: string
          example: Av. Siempre Viva 742
        distrito:
          type: string
          example: Miraflores
        provincia:
          type: string
          example: Lima
        departamento:
          type: string
          example: Lima
        pais:
          type: string
          example: Perú

    # 📌 Catálogos
...